fix compilation on Darwin / Mac

+++ lsplatform/mac_device.go
@@ -1,4 +1,4 @@
-//go:build aix || dragonfly || freebsd || linux || netbsd || openbsd || zos
+//go:build darwin
 
 package lsplatform
 
@@ -11,7 +11,7 @@ import (
 
 func (*LocalPlatform) DeviceNumbers(info FileInfo) (string, error) {
 	stat := info.Sys().(*syscall.Stat_t)
-	major := strconv.FormatInt(int64(unix.Major(stat.Rdev)), 10)
-	minor := strconv.FormatInt(int64(unix.Minor(stat.Rdev)), 10)
+	major := strconv.FormatInt(int64(unix.Major(uint64(stat.Rdev))), 10)
+	minor := strconv.FormatInt(int64(unix.Minor(uint64(stat.Rdev))), 10)
 	return major + "," + minor, nil
 }
